managed to add widgets that correspond to an overwrite button and a delete button in the top right panel. Need to connect those to methods that would overwrite and delete selected annotations todo: (i)   write a method that overwrites the selected annotation (ii)  link above method to overwrite annotation (iii) do the same for polylines and rectItem
added a filedialog that checks with the user if this is ok to delete stuff. The filedialog is initialised inside AnnotationModelOptions, hence need to have a qline edit that displays the file loc of the annotation being viewed. hence need to re write the initialisation of this file

need to also compuet the max length of things in the dropdown and set it to be this